Retry request upon rate limit exceeded
This change adds the ability to retry a failed request
that received a 429 (Too Many Requests) response. This
occurs when the Twitch request rate limit is exhausted
for the provided client ID.

This feature is enabled by default if a `RateLimitFunc`
is provided, but can be disabled by setting the
`RetryRateLimitedRequests` option to false.