Attach app access token when user assess token is not set
As reported in #3, it is possible to use an app access token to
extend the request rate limit to 120. Previously it was thought
that only user access tokens were able to do this. Unfortunately
this is not documented in Twitch's API docs.

This implementation sees that app access tokens be set as the
bearer token if an access token is not provided. If both app
and user access tokens are provided, user tokens are given priority.
Meaning that the app access token will be ignored.

Closes #3